+++ eds/blocks/hero/hero.js
@@ -1,4 +1,6 @@
 import { domEl } from '../../scripts/dom-helpers.js';
+import { decorateToCalciteBtn } from '../../scripts/scripts.js';
+
 
 export default function decorate(block) {
   block.querySelectorAll('picture > img').forEach((img, i) => {
@@ -15,10 +17,12 @@ export default function decorate(block) {
   videoElement.toggleAttribute('autoplay', true);
   if (videoAssets.length === 2) {
     videoSrc.setAttribute('src', videoAssets[1].getAttribute('title'));
-    videoAssets[1].classList.add('video-src');
+    videoAssets[1].classList.add('hidden');
   } else {
-    videoSrc.setAttribute('src', videoAssets[0].getAttribute('title'));
-    videoAssets[0].classList.add('video-src');
+    if (videoAssets[0] !== undefined) {
+      videoSrc.setAttribute('src', videoAssets[0].getAttribute('title'));
+      videoAssets[0].classList.add('hidden');
+    }
   }
   videoSrc.setAttribute('type', 'video/mp4');
 
@@ -43,4 +47,5 @@ export default function decorate(block) {
     }
   });
   contentDiv.remove(); // Remove the empty div
+  decorateToCalciteBtn(block);
 }
