+++ eds/blocks/storyteller/storyteller.js
@@ -1,3 +1,5 @@
+import { decorateToCalciteBtn } from '../../scripts/scripts.js';
+
 /**
  * Determine if mp4 resource is available. Add selector 'foreground-container' to p tag.
  * @param {string} vidUrls array with href property
@@ -103,27 +105,6 @@ function getMP4(block) {
   return mp4Urls;
 }
 
-/**
- * convert link button to calcite button
- * @param {element} block The block element
- */
-function decorateLinkBtn(block) {
-  const anchorElements = block.querySelectorAll('a');
-  anchorElements.forEach((anchorElement) => {
-    if (!anchorElement.classList.contains('hidden')) {
-      const calciteButton = document.createElement('calcite-button');
-      calciteButton.innerHTML = anchorElement.innerHTML;
-      if (anchorElement.getAttribute('href')) {
-        calciteButton.setAttribute('href', anchorElement.getAttribute('href'));
-      }
-      if (!calciteButton.getAttribute('scale')) {
-        calciteButton.setAttribute('scale', 'l');
-      }
-      anchorElement.replaceWith(calciteButton);
-    }
-  });
-}
-
 export default async function decorate(block) {
   const pTags = block.querySelectorAll('p');
   const pictureTagLeft = pTags[0].querySelector('picture');
@@ -184,5 +165,5 @@ export default async function decorate(block) {
     toggleVideo(videoBtn);
   });
 
-  decorateLinkBtn(block);
+  decorateToCalciteBtn(block);
 }
